requirements.txt
python-telegram-bot==21.4
openai>=1.100.0,<2
